Lots of new things :D
Logging changed. Now logs to db3 as well

Spawns are now force bound (regardless of who they are spawned for)

Online players now lists ALL THE THINGS. Current things to list: Online Players, Players in town, in a dungeon, on a boat, fishing, in fruit fight, in Arena, in Monster Brawl, in Capture the relic, Party members of currently running dungeons and with plans to list details about those running dungeons like the map and the time spent in dungeon so far.